+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Events section to member profile, showing upcoming events the person is attending
+
 ## [2.2.2] - 2020-12-09
 ### Added
 - Added Skills to Learn to user profiles
@@ -13,7 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Couldnt click Edit Profile button
 - Bug with routing when clicking on a project from a user profile and then trying to close it
-
 
 ## [2.2.1] - 2020-12-05
 ### Added